add checkpointbyUUIDS
