+++ master/internal/core_experiment.go
@@ -85,10 +85,10 @@ func (m *Master) getExperimentCheckpointsToGC(c echo.Context) (interface{}, erro
 	}
 	// We need to write a db method to get all checkpoints objects for the given list of checkpoints uuids.
 	// Use checkpoints_view for this. This needs to be a json object. Look at CheckpointByUUID in postgres_trial.go.
-	// You can then marshal the model.Checkpoint struct.
+	// You can then marshal the model.Checkpoint struct. So that experiment.py can access it properly.
 
-	// So that experiment.py can access it properly.
 	// Get metric name for experiments from experiment table (see how they extract it from GCRaw method.)
+
 	// Ssh in into latest master and use the GET checkpoints query (from checkpoints_view) and see how it performs.
 	return m.db.ExperimentCheckpointsToGCRaw(
 		args.ExperimentID, args.ExperimentBest, args.TrialBest, args.TrialLatest)
